+++ eng/scripts/automation/Invoke-DataPlaneGenerateSDKPackage.ps1
@@ -4,6 +4,7 @@ param (
   [string]$namespace,
   [string]$sdkPath,
   [string]$inputfiles="", #input files, separated by semicolon if more than one
+  [stirng]$readme = "",
   [string]$securityScope="",
   [string]$securityHeaderName="",
   [string]$AUTOREST_CONFIG_FILE = "autorest.md"
@@ -12,7 +13,7 @@ param (
 
 # Generate dataplane library
 $outputJsonFile = "newpackageoutput.json"
-New-DataPlanePackageFolder -service $service -namespace $namespace -sdkPath $sdkPath -inputfiles $inputfiles -securityScope $securityScope -securityHeaderName $securityHeaderName -AUTOREST_CONFIG_FILE $AUTOREST_CONFIG_FILE -outputJsonFile $outputJsonFile
+New-DataPlanePackageFolder -service $service -namespace $namespace -sdkPath $sdkPath -inputfiles $inputfiles -readme $readme -securityScope $securityScope -securityHeaderName $securityHeaderName -AUTOREST_CONFIG_FILE $AUTOREST_CONFIG_FILE -outputJsonFile $outputJsonFile
 if ( $? -ne $True) {
   Write-Error "Failed to create sdk project folder. exit code: $?"
   exit 1